+++ src/main/kotlin/minesweeper/domain/CellType.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 enum class CellType {
     NONE,
