+++ src/test/kotlin/minesweeper/domain/ColumnTest.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
